Struct zaplib::Padding

pub struct Padding {
    pub l: f32,
    pub t: f32,
    pub r: f32,
    pub b: f32,
}

Inner padding dimensions that should be applied on top of a Rect or other object that defines dimensions.

TODO(JP): these values can be negative, which can be quite confusing, but we seem to actually honor that in the layout boxes code. Might be good to look into that and see if we should forbid that or not (we seem to never actually do that yet).

Fields
l: f32t: f32r: f32b: f32
Implementations

impl Padding

pub const ZERO: Padding

pub const DEFAULT: Padding

TODO(JP): Replace these with Padding::default() when https://github.com/rust-lang/rust/issues/67792 gets done

pub const fn all(v: f32) -> Padding

pub const fn left(v: f32) -> Padding

pub const fn top(v: f32) -> Padding

pub const fn right(v: f32) -> Padding

pub const fn bottom(v: f32) -> Padding

pub const fn vh(v: f32, h: f32) -> Padding

Helper function to set vertical and horizontal padding This is a common case when top=bottom left=right