Struct zaplib::TextIns

#[repr(C)]
pub struct TextIns {
    pub font_t1: Vec2,
    pub font_t2: Vec2,
    pub color: Vec4,
    pub rect_pos: Vec2,
    pub rect_size: Vec2,
    pub char_depth: f32,
    pub base: Vec2,
    pub font_size: f32,
    pub char_offset: f32,
    pub marker: f32,
}

Fields
font_t1: Vec2

Texture coordinates for the bottom-left corner of the glyph in the texture atlas

font_t2: Vec2

Texture coordinates for the top-right corner of the glyph in the texture atlas

color: Vec4

Color for a glyph, usually set at the same color as TextIns

rect_pos: Vec2

Glyph position in view space

rect_size: Vec2

Glyph size in view space

char_depth: f32

Depth offset (prevents z-fighting)

base: Vec2

Position used in TextIns::closest_offset.

font_size: f32

Font size in pixels

char_offset: f32

Character index in the text string

marker: f32

TODO(JP): document.

Implementations

impl TextIns

pub fn generate_2d_glyphs<F>(
text_style: &TextStyle,
fonts_data: &RwLock<CxFontsData>,
dpi_factor: f32,
font_scale: f32,
draw_depth: f32,
color: Vec4,
pos: Vec2,
char_offset: usize,
chunk: &[char],
char_callback: F
) -> Vec<TextIns> where
F: FnMut(char, usize, f32, f32) -> f32,

pub fn set_color(cx: &mut Cx, area: Area, color: Vec4)

pub fn draw_glyphs(
cx: &mut Cx,
glyphs: &[TextIns],
props: &DrawGlyphsProps
) -> Area

pub fn draw_glyphs_with_scroll_sticky(
cx: &mut Cx,
glyphs: &[TextIns],
text_style: &TextStyle,
horizontal: bool,
vertical: bool
) -> Area

pub fn draw_str(
cx: &mut Cx,
text: &str,
pos: Vec2,
props: &TextInsProps
) -> Area

pub fn draw_walk(cx: &mut Cx, text: &str, props: &TextInsProps) -> Area

TODO(JP): This doesn’t seem to work well with Direction::Down (or other directions for that matter). Not a high priority but might good to be aware of.

TextInsProps::position_anchoring is ignored by this function.

pub fn closest_offset(
cx: &Cx,
area: &Area,
pos: Vec2,
line_spacing: f32
) -> Option<usize>

Looks up text with the behavior of a text selection mouse cursor.

pub fn get_monospace_base(cx: &Cx, text_style: &TextStyle) -> Vec2