Struct zaplib::Window

pub struct Window {
    pub window_id: Option<usize>,
    pub create_inner_size: Option<Vec2>,
    pub create_position: Option<Vec2>,
    pub create_title: String,
    pub create_add_drop_target_for_app_open_files: bool,
}

A pointer to a [CxWindow] (indexed in Cx::windows using Window::window_id),

Fields
window_id: Option<usize>

The id referring to [CxWindow], once instantiated. It’s an index in Cx::windows.

create_inner_size: Option<Vec2>

The inner dimensions of the native window when it’s created for the first time.

create_position: Option<Vec2>

The position on the screen of the native window when it’s created for the first time.

create_title: String

The title that the window will get once created.

create_add_drop_target_for_app_open_files: bool

Set to true if the window should as a drop target for a AppOpenFiles events.

TODO(JP): only works on the wasm32 and mac targets for now.

Implementations

impl Window

pub fn begin_window(&mut self, cx: &mut Cx)

pub fn get_inner_size(&mut self, cx: &mut Cx) -> Vec2

pub fn get_position(&mut self, cx: &mut Cx) -> Option<Vec2>

pub fn set_position(&mut self, cx: &mut Cx, pos: Vec2)

pub fn handle_window(&mut self, _cx: &mut Cx, _event: &mut Event) -> bool

pub fn end_window(&mut self, cx: &mut Cx) -> Area

pub fn minimize_window(&mut self, cx: &mut Cx)

pub fn maximize_window(&mut self, cx: &mut Cx)

pub fn fullscreen_window(&mut self, cx: &mut Cx)

pub fn normal_window(&mut self, cx: &mut Cx)

pub fn can_fullscreen(&mut self, cx: &mut Cx) -> bool

pub fn xr_can_present(&mut self, cx: &mut Cx) -> bool

pub fn is_fullscreen(&mut self, cx: &mut Cx) -> bool

pub fn xr_is_presenting(&mut self, cx: &mut Cx) -> bool

pub fn xr_start_presenting(&mut self, cx: &mut Cx)

pub fn xr_stop_presenting(&mut self, cx: &mut Cx)

pub fn is_topmost(&mut self, cx: &mut Cx) -> bool

pub fn set_topmost(&mut self, cx: &mut Cx, topmost: bool)

pub fn restore_window(&mut self, cx: &mut Cx)

pub fn close_window(&mut self, cx: &mut Cx)